move getStatuses from membership to groups
