+++ src/services/groups/service/groups.ts
@@ -152,3 +152,29 @@ export const update = function (
     payload: OpenAPI.operations['update_group_v2_groups__group_id__put']['requestBody']['content']['application/json'];
   }
 >;
+
+/**
+ * Get the status counts of memberships for each group you are an admin or manager of.
+ * @see https://groups.api.globus.org/redoc#tag/groups/operation/get_statuses_v2_groups_statuses_get
+ */
+export const getStatuses = function (
+  options?,
+  sdkOptions?,
+): Promise<
+  JSONFetchResponse<
+    OpenAPI.operations['get_statuses_v2_groups_statuses_get']['responses']['200']['content']['application/json']
+  >
+> {
+  return serviceRequest(
+    {
+      service: ID,
+      resource_server: RESOURCE_SERVERS.GROUPS,
+      path: `/v2/groups/statuses`,
+    },
+    options,
+    sdkOptions,
+  );
+} satisfies ServiceMethod<{
+  query?: never;
+  payload?: never;
+}>;
