+++ src/services/groups/service/membership.ts
@@ -3,11 +3,7 @@ import { HTTP_METHODS, serviceRequest } from '../../shared.js';
 import { RESOURCE_SERVERS } from '../../auth/config.js';
 
 import type { OpenAPI } from '../index.js';
-import type {
-  JSONFetchResponse,
-  ServiceMethod,
-  ServiceMethodDynamicSegments,
-} from '../../types.js';
+import type { JSONFetchResponse, ServiceMethodDynamicSegments } from '../../types.js';
 
 /**
  * Perform actions on members of the group.
@@ -39,29 +35,3 @@ export const act = function (
     payload: OpenAPI.operations['group_membership_post_actions_v2_groups__group_id__post']['requestBody']['content']['application/json'];
   }
 >;
-
-/**
- * Get the status counts of memberships for each group you are an admin or manager of.
- * @see https://groups.api.globus.org/redoc#tag/groups/operation/get_statuses_v2_groups_statuses_get
- */
-export const getStatuses = function (
-  options?,
-  sdkOptions?,
-): Promise<
-  JSONFetchResponse<
-    OpenAPI.operations['get_statuses_v2_groups_statuses_get']['responses']['200']['content']['application/json']
-  >
-> {
-  return serviceRequest(
-    {
-      service: ID,
-      resource_server: RESOURCE_SERVERS.GROUPS,
-      path: `/v2/groups/statuses`,
-    },
-    options,
-    sdkOptions,
-  );
-} satisfies ServiceMethod<{
-  query?: never;
-  payload?: never;
-}>;
